chore: change employee serial number type to int

+++ src/main/java/org/entreprise/employee/Employee.java
@@ -13,7 +13,7 @@ import java.time.LocalDate;
 public class Employee {
     private String firstName;
     private String lastName;
-    private Long serialNumber;
+    private int serialNumber;
     private LocalDate dateDeNaissance;
     private LocalDate hiringDate;
     private LocalDate contractEndDate;
